+++ contracts/interfaces/IRelayERC20.sol
@@ -4,8 +4,7 @@ pragma solidity 0.8.4;
 import "../external/zeppelin/token/ERC20/IERC20.sol";
 
 interface IRelayERC20 is IERC20 {
-
-    function relayNonce(address account) external returns (uint);
+    function relayNonce(address account) external returns (uint256);
 
     function relayedTransfer(
         bytes calldata sig,
@@ -14,4 +13,4 @@ interface IRelayERC20 is IERC20 {
         uint256 amount,
         uint256 fee
     ) external;
-} 
+}
